Utilisation de la librairie d'affichage fournie par le formateur, ajout de classe de service, corrections de problèmes de mapping

+++ BoVoyage.Core/Entites/Voyage.cs
@@ -22,9 +22,8 @@ namespace BoVoyage.Core.Entites
         public DateTime DateAller { get; set; }
         public DateTime DateRetour { get; set; }
         public int PlacesDisponibles { get; set; }
-        public double PrixParPersonne { get; set; }
+        public decimal PrixParPersonne { get; set; }
 
         public static void Reserver() { }
-
     }
 }
